kitchen requests

+++ database/migrations/2021_04_27_193334_create_warehouse_stocks_table.php
@@ -17,7 +17,7 @@ class CreateWarehouseStocksTable extends Migration
             $table->id();
             $table->bigInteger('material_id')->unsigned();
                 $table->foreign('material_id')->references('id')->on('materials');
-            $table->integer('quantity')->unsigned();
+            $table->decimal('quantity')->unsigned();
             $table->softDeletes();
             $table->timestamps();
         });
